rename _api_funcs.py to _convert.py
