+++ src/h3/api/basic_str/_convert.py
